Move all image preview code into reusable component
